feat: moved cache reader to grpc service
