+++ internal/grpc/cache.go
@@ -1,4 +1,4 @@
-package api
+package grpc
 
 import (
 	"context"
